Began implementing A* search functions

+++ src/Entity.cpp
@@ -1,12 +1,12 @@
 #include "Entity.h"
 
-Entity::Entity(int id, int x, int y)
+Entity::Entity(int id, float x, float y)
 {
     _id = id;
     this->setLocation(x, y);
 }
 
-void Entity::setLocation(int x, int y)
+void Entity::setLocation(float x, float y)
 {
     _x = x;
     _y = y;
